requirements.txt
behave==1.2.6
etlhelper==0.9.6
parse==1.18.0
parse-type==0.5.2
PyHamcrest==2.0.2
six==1.15.0
allure-behave==2.8.31
SQLAlchemy==1.3.22
pymysql==1.0.2
mysqldbmodel==1.1.0
numpy==1.19.5
